+++ Pets/Pets.Api/Controllers/AnimalsController.cs
@@ -1,7 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Pets.Infrastructure.Commands.Animals;
 using Pets.Infrastructure.DTO;
 using Pets.Infrastructure.Services;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -37,28 +39,56 @@ namespace Pets.Api.Controllers
             return await _animalService.GetAllAsync(email);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> Post(string email, [FromBody]CreateAnimal request)
         {
+            if (email != await GetLoggedUserEmail())
+            {
+                throw new Exception("You can only add animals to your account.");
+            }
+
             await _animalService.AddAsync(email, request.Name, request.YearOfBirth);
 
             return NoContent();
         }
 
+        [Authorize]
         [HttpPut("{name}")]
         public async Task<IActionResult> Put(string email, string name, [FromBody]UpdateAnimal request)
         {
+            if (email != await GetLoggedUserEmail())
+            {
+                throw new Exception("You can only edit animals in your account.");
+            }
+
             await _animalService.UpdateAsync(email, name, request.Name, request.YearOfBirth);
 
             return NoContent();
         }
 
+        [Authorize]
         [HttpDelete("{name}")]
         public async Task<IActionResult> Delete(string email, string name)
         {
+            if (email != await GetLoggedUserEmail())
+            {
+                throw new Exception("You can only delete animals in your account.");
+            }
+
             await _animalService.DeleteAsync(email, name);
 
             return NoContent();
+        }
+
+        public async Task<string> GetLoggedUserEmail()
+        {
+            if (HttpContext.User.Identity.Name == null)
+            {
+                throw new System.Exception("There is no logged user.");
+            }
+
+            return await Task.FromResult(HttpContext.User.Identity.Name);
         }
     }
 }
